Handles the new JSON format from the API
The API previously returned an object of lists, whereas it now returns a
list of objects, meaning a subtly different parse is required.
This returns a slightly different data frame to what was previously
returned, although I can't say for sure how different as I can't access
the old API to get a benchmark.

One difference is that both window times are no longer available,
instead only a single timestamp is returned corresponding to the
specified time averaging convention.