Update get_envirologger_data to handle new API response
The response from the API is now nested into 1 min elements, so the parsing of this is now handled within a `map_dfr()` call.

It also seems to no longer report date and date_end, rather just one date and the type of date, so this is returned instead.

The way the PreScaled and Scaled tables are returned required unnesting, so added `tidyr` to imports.

Finally added the new purrr progress bar as an option, true by default. Can be toggled with the `.progress` argument.